minor changes to finish the project
nothing important, just translating some stuff

+++ view/frontend/adminTools/editArticles.php
@@ -1,7 +1,7 @@
 <?php
 
   
-session_start();
+
 
 if (isset($_SESSION['login']) && isset($_SESSION['password'])) { ?>
 
@@ -45,11 +45,11 @@ The above copyright notice and this permission notice shall be included in all c
       Tip 2: you can also add an image using data-image tag
        -->
       <div class="logo">
-        <a href="http://www.creative-tim.com" class="simple-text logo-mini">
-          CT
+        <a href="" class="simple-text logo-mini">
+          Tableau
         </a>
-        <a href="http://www.creative-tim.com" class="simple-text logo-normal">
-          Creative Tim
+        <a href="" class="simple-text logo-normal">
+          de bord
         </a>
       </div>
       <div class="sidebar-wrapper">
@@ -90,7 +90,7 @@ The above copyright notice and this permission notice shall be included in all c
       <nav class="navbar navbar-expand-lg navbar-transparent navbar-absolute fixed-top ">
         <div class="container-fluid">
           <div class="navbar-wrapper">
-            <a class="navbar-brand" href="javascript:;">test</a>
+            <a class="navbar-brand" href="javascript:;">Tableau de bord des articles</a>
           </div>
           <button class="navbar-toggler" type="button" data-toggle="collapse" aria-controls="navigation-index" aria-expanded="false" aria-label="Toggle navigation">
             <span class="sr-only">Toggle navigation</span>
@@ -99,40 +99,10 @@ The above copyright notice and this permission notice shall be included in all c
             <span class="navbar-toggler-icon icon-bar"></span>
           </button>
           <div class="collapse navbar-collapse justify-content-end">
-            <form class="navbar-form">
-              <div class="input-group no-border">
-                <input type="text" value="" class="form-control" placeholder="Search...">
-                <button type="submit" class="btn btn-white btn-round btn-just-icon">
-                  <i class="material-icons">search</i>
-                  <div class="ripple-container"></div>
-                </button>
-              </div>
-            </form>
+
             <ul class="navbar-nav">
-              <li class="nav-item">
-                <a class="nav-link" href="javascript:;">
-                  <i class="material-icons">dashboard</i>
-                  <p class="d-lg-none d-md-block">
-                    Stats
-                  </p>
-                </a>
-              </li>
-              <li class="nav-item dropdown">
-                <a class="nav-link" href="http://example.com" id="navbarDropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                  <i class="material-icons">notifications</i>
-                  <span class="notification">5</span>
-                  <p class="d-lg-none d-md-block">
-                    Some Actions
-                  </p>
-                </a>
-                <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdownMenuLink">
-                  <a class="dropdown-item" href="#">Mike John responded to your email</a>
-                  <a class="dropdown-item" href="#">You have 5 new tasks</a>
-                  <a class="dropdown-item" href="#">You're now friend with Andrew</a>
-                  <a class="dropdown-item" href="#">Another Notification</a>
-                  <a class="dropdown-item" href="#">Another One</a>
-                </div>
-              </li>
+
+
               <li class="nav-item dropdown">
                 <a class="nav-link" href="javascript:;" id="navbarDropdownProfile" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                   <i class="material-icons">person</i>
@@ -176,9 +146,11 @@ The above copyright notice and this permission notice shall be included in all c
                       <hr>
                       <!-- Pager -->
 
-                      <div class="clearfix">                        
-                        <a class="btn btn-primary float-right" href="posts.php?action=editPosts&amp;id=<?=$data['id']?>">Editer cet article &rarr;</a>
-                        <a class="btn btn-primary float-right" href="posts.php?action=delete&amp;id=<?=$data['id']?>">Supprimer l'article &rarr;</a>                     
+                      <div class="clearfix">                
+
+                        <a class="btn btn-primary float-right" href="posts.php?action=delete&amp;id=<?=$data['id']?>" onclick="return confirm('ëtes-vous sur de vouloir supprimer ce commentaire ?');">Supprimer l'article &rarr;</a>
+                        <a class="btn btn-primary float-right" href="posts.php?action=editPosts&amp;id=<?=$data['id']?>">Editer l'article &rarr;</a>
+                    
                       </div>
                     </div>
                   </div>
@@ -192,34 +164,10 @@ The above copyright notice and this permission notice shall be included in all c
         <div class="container-fluid">
           <nav class="float-left">
             <ul>
-              <li>
-                <a href="https://www.creative-tim.com">
-                  Creative Tim
-                </a>
-              </li>
-              <li>
-                <a href="https://creative-tim.com/presentation">
-                  About Us
-                </a>
-              </li>
-              <li>
-                <a href="http://blog.creative-tim.com">
-                  Blog
-                </a>
-              </li>
-              <li>
-                <a href="https://www.creative-tim.com/license">
-                  Licenses
-                </a>
-              </li>
             </ul>
           </nav>
           <div class="copyright float-right">
-            &copy;
-            <script>
-              document.write(new Date().getFullYear())
-            </script>, made with <i class="material-icons">favorite</i> by
-            <a href="https://www.creative-tim.com" target="_blank">Creative Tim</a> for a better web.
+
           </div>
         </div>
       </footer>
@@ -227,9 +175,7 @@ The above copyright notice and this permission notice shall be included in all c
   </div>
   <div class="fixed-plugin">
     <div class="dropdown show-dropdown">
-      <a href="#" data-toggle="dropdown">
-        <i class="fa fa-cog fa-2x"> </i>
-      </a>
+
       <ul class="dropdown-menu">
         <li class="header-title"> Sidebar Filters</li>
         <li class="adjustments-line">
